refactor(config): remove plugins, module
